Add files via upload
Got some of the bottom portion done.  Codes pretty gross.  I will fix it up later.  Trying to figure out the best way to do the error line.  If we put it in the left vbox it will get out of order when the combo box order changes I think.  Still need to figure out the best way to keep updated information in the drop-downs too, but we can figure that out later.