Add files via upload
I got the monthly report working and delete function working.  I still wasn't able to figure out the x-axis sorting.  Once/if we do figure it out it should be easy to apply to all report types that require it.  Date Range report is partially implemented.  I have it pretty much all coded in GUI but, still need to do some implemented methods in the handler for it to work, unsure of how to do implement those.  I also started to implement the modify/add data feature.  I changed some UI for that and have added some methods to handler for it so far.